Add doShowMessageCount.js
