+++ app/scripts.babel/main.js
@@ -17,6 +17,7 @@ import doChangeThumbnailSize from '@actions/doChangeThumbnailSize';
 import doConfirmSendMessageButton from '@actions/doConfirmSendMessageButton';
 import doExpandUserIcon from '@actions/doExpandUserIcon';
 import doResponsiveMultiView from '@actions/doResponsiveMultiView';
+import doShowMessageCount from '@actions/doShowMessageCount';
 
 import FileTypes from '@enums/FileTypes';
 import FormTypes from '@enums/FormTypes';
@@ -377,33 +378,6 @@ import settingData from '@constants/settingData';
       .filter(key => settings[key] === true)
       .map(key => SETTINGS_KEY_ACTIONS[key])
       .forEach(action => action());
-  }
-
-  /**
-  * 入力文字数の表示機能を実行します。
-  */
-  async function doShowMessageCount(){
-    const settings = await fetchSettings();
-    const countDown = settings.show_message_count_types== 'countdown';
-
-    const sendForms = $('.form-send');
-    sendForms.each((i, sendForm) => {
-      const $textArea = $(sendForm).find('.form-send-text');
-      const maxLength = $textArea.prop('maxLength');
-
-      //カウンターを作成
-      const count = countDown ? maxLength : 0;
-      const $counter = $(`<label>${count}</label>`).css('margin-right', '8px');
-      const $sendButtonGroup = $(sendForm).find('.form-send-button-group');
-      $sendButtonGroup.prepend($counter);
-
-      //文字入力時にカウンターの値を更新
-      $textArea.on('input.direct_helper_doShowMessageCount', () => {
-        const currentLength = $textArea.val().length;
-        const count = countDown ? maxLength - currentLength : currentLength;
-        $counter.text(count);
-      });
-    });
   }
 
   /**
